Read number of input channels directly (#8)
This commit adjusts Paradis.__init__ to read the number of
input channels directly from
datamodule.dataset.num_in_features.  Previously, this was
computed by adding the dynamic input features (in dataset)
and the number of constant features (in config).

This appears to have missed the 'forcings' variables,
causing a fresh install to immediately crash because it
sees 97 channels when it expected 92.

The fix to use num_in_features avoids the problem of
multiple sources of authority, and it should be robust to
future changes in the dataset code.

Co-authored-by: Christopher Subich <Christopher.Subich@canada.ca>