Add dataset preload option
This commit adds 'preload' options to the dataset
configuration sections.  If enabled, the data will be
preloaded from disk when the Era5DataModule is created,
storing everything in memory.

This will take a very large amount of system RAM depending
on the model resolution, but enabling this for
coarse-resolution (5.625°) training runs approximately
doubles the inference speed.

This commit also sets the DataLoader workers as persistent,
elminating 'hitches' when validation or a new training epoch
begins.  Without persistent workers, the processes are
created anew on these events.