Quality of life (#9)
* Add configuration options for seeding, validation

This commit adds the following configuration options:

* init.seed: sets the random seed for the model
* training.max_steps: sets the maximum number of steps for
  training, rather than a maximum number of data epochs
* training.dataset.validation_every_n_steps: run validation
  every N steps rather than once per epoch
* training.dataset.validation_batches: If not null, use N
  batches only for validation.  If set to 0, disables
  validation.

* Extend wsd schedule to allow specifying # of warmup/decay steps

This commit replaces `training.scheduler.wsd.warmup_pct` and
`decay_pct` with `warmup` and `decay`.  Values less than 1
are treated as fractions (as `_pct` were), and otherwise the
value refers to a fixed number of steps.

* Add dataset preload option

This commit adds 'preload' options to the dataset
configuration sections.  If enabled, the data will be
preloaded from disk when the Era5DataModule is created,
storing everything in memory.

This will take a very large amount of system RAM depending
on the model resolution, but enabling this for
coarse-resolution (5.625°) training runs approximately
doubles the inference speed.

This commit also sets the DataLoader workers as persistent,
elminating 'hitches' when validation or a new training epoch
begins.  Without persistent workers, the processes are
created anew on these events.

* Disabling checkpoints in config properly disables checkpoints

* Remove v_num item from progress bar

---------

Co-authored-by: Christopher Subich <Christopher.Subich@canada.ca>