Replace nn.Modules in paradis_multilayer with equivalent-configuration GMBlocks (#11)
* Fix number of dynamic channels

* Formatting

* Read number of input channels directly (#8)

This commit adjusts Paradis.__init__ to read the number of
input channels directly from
datamodule.dataset.num_in_features.  Previously, this was
computed by adding the dynamic input features (in dataset)
and the number of constant features (in config).

This appears to have missed the 'forcings' variables,
causing a fresh install to immediately crash because it
sees 97 channels when it expected 92.

The fix to use num_in_features avoids the problem of
multiple sources of authority, and it should be robust to
future changes in the dataset code.

Co-authored-by: Christopher Subich <Christopher.Subich@canada.ca>

* Quality of life (#9)

* Add configuration options for seeding, validation

This commit adds the following configuration options:

* init.seed: sets the random seed for the model
* training.max_steps: sets the maximum number of steps for
  training, rather than a maximum number of data epochs
* training.dataset.validation_every_n_steps: run validation
  every N steps rather than once per epoch
* training.dataset.validation_batches: If not null, use N
  batches only for validation.  If set to 0, disables
  validation.

* Extend wsd schedule to allow specifying # of warmup/decay steps

This commit replaces `training.scheduler.wsd.warmup_pct` and
`decay_pct` with `warmup` and `decay`.  Values less than 1
are treated as fractions (as `_pct` were), and otherwise the
value refers to a fixed number of steps.

* Add dataset preload option

This commit adds 'preload' options to the dataset
configuration sections.  If enabled, the data will be
preloaded from disk when the Era5DataModule is created,
storing everything in memory.

This will take a very large amount of system RAM depending
on the model resolution, but enabling this for
coarse-resolution (5.625°) training runs approximately
doubles the inference speed.

This commit also sets the DataLoader workers as persistent,
elminating 'hitches' when validation or a new training epoch
begins.  Without persistent workers, the processes are
created anew on these events.

* Disabling checkpoints in config properly disables checkpoints

* Remove v_num item from progress bar

---------

Co-authored-by: Christopher Subich <Christopher.Subich@canada.ca>

* Consolidate nn.module use into simple and GMBlock (#10)

* Handle checkpoint directly with lightning

* Fix error message if no/wrong LR schedule enabled

* Log gradient magnitude

* Log training step time

* Display training progress bar over all training steps

This commit modifies the training progress bar to one that
is continuous over the whole training process.  Previously,
Lightning closed and re-opened a bar for each epoch, which
made estimation of the total remaining time/etc difficult
and manual.

* train.py: also disable progress bar if progress bar disabled

Previously, train.py would try to create a progress bar if
per-step loss-printing was disabled.  This would cause a
conflict if cfg.training.progress_bar was also False, since
the latter property is used in the callback-creation.

* Progress bar: preserve validation bar in jupyter notebooks

This commit modifies utils/callbacks.py to preserve the
validation progress bar when running inside a jupyter
notebook.  Without this, the validation bar is opened/closed
for each validation stage, and the associated events cause
the jupyter notebook to think that the cell's output is
filling up.

* Type annotation fixes

* Adjust ActivationType (CLPBlock) for simplicity

* Fix type-checking warnings in Era5Dataset

* Add simple_blocks

This commit adds model/simple_blocks.py, which consolidates
the linear and near-linear layers used in Paradis into a
single interface for later plug-and-play interchangeability.
Blocks currently implemented are:

(The convolutional blocks with nontrivial kernel also
automatically call GeoCyclicPadding)

* FullConv: implements a "full" convolution, allowing
interactions across channels between neighbouring grid cells
* FlatConv: Implements a "flat" convolution that operates
channelwise
* CLinear: Implements a pointwise linear transform across
the channel dimension (equivalent to nn.Linear)
* SepConv: A FlatConv followed by CLinear, allowing some
channel interactions with fewer parameters (C^2 + C·k^2,
rather than C^2 k^2)

* GlobalNorm: Global normalization via nn.LayerNorm, setting
the global mean to 0 and the global standard deviation to 1,
modified by learnable scale and bias parameters (n.b.:
2·NLat·NLon·NChannel parameters!)
* ChannelNorm: Pointwise normalization across the channel
dimension only (2*NChannel learnable parameters)

* NormedConv: Implements a FullConv followed by a
GlobalNorm, matching the current behaviour of the first
stage of CLPBlock

* GlobalBias: Learns N (user-specified) bias channels per
gridpoint along with a linear projection to C output
channels (N·Lat·Lon + N·C parameters)

* Add GMBlock, Generic Multilayer Block

This commit adds model/gmblock.py and the associated GMBlock
class.  This class acts as a generic multilayer block,
capable of stacking various simple blocks
(model/simple_blocks.py) with enough configuration options
to fully replicate all existing uses of CLPBlock.  The
ultimate intent is to expose the layer settings at runtime /
in the configuration file, allowing semi-automated searches
of configuration space.

---------

Co-authored-by: Carlos Pereira <carlosp28@gmail.com>
Co-authored-by: Christopher Subich <Christopher.Subich@canada.ca>

* Type annotation fixes in paradis_multilayer

* Replace CLP and other nn.Modules in paradis_multilayer with GMBlock

As an example of how to use GMBlock, this commit replaces
all of the direct invocations of nn.Modules and CLP inside
paradis_multilayer with equivalent GMBlock constructions.

This commit also re-enables the semi-Lagrangian advection
module (rather than double-convolution), so please be
careful when applying this commit.

---------

Co-authored-by: Carlos Pereira <carlosp28@gmail.com>
Co-authored-by: Christopher Subich <Christopher.Subich@canada.ca>